Break out cube_routes into different files.
Draft and edit routes now have their own files.
